+++ web/src/nav/GlobalNavbar.test.tsx
@@ -43,6 +43,7 @@ const PROPS: React.ComponentProps<typeof GlobalNavbar> = {
     variant: 'default',
     globbing: false,
     showOnboardingTour: false,
+    branding: undefined,
 }
 
 describe('GlobalNavbar', () => {
@@ -53,9 +54,6 @@ describe('GlobalNavbar', () => {
 
     test('low-profile', () =>
         expect(renderer.create(<GlobalNavbar {...PROPS} variant="low-profile" />).toJSON()).toMatchSnapshot())
-
-    test('low-profile-with-logo', () =>
-        expect(renderer.create(<GlobalNavbar {...PROPS} variant="low-profile-with-logo" />).toJSON()).toMatchSnapshot())
 
     test('no-search-input', () =>
         expect(renderer.create(<GlobalNavbar {...PROPS} variant="no-search-input" />).toJSON()).toMatchSnapshot())